Use WebAssembly to parse local .bam files in the browser (#20)
To parse local .bam files, Ribbon currently relies on a JavaScript rewrite of portions of samtools (via bam.iobio and biodalliance). This PR replaces that JavaScript code with samtools directly, by first compiling it to WebAssembly. It also uses Aioli to run samtools.wasm inside a WebWorker (i.e. a separate thread), so that the main thread used for the UI is not disrupted.

This PR does not replace the reliance on bam.iobio for parsing remote .bam files from URLs. Although it is easy to do that from the browser directly, it would only work with URLs that are CORS-enabled, which most are not.